feat(20433): add translation, files renaming, key navigation enhancements
